% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_clustering.R
\name{auto_clustering}
\alias{auto_clustering}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
auto_clustering(data, x, y)
}
\arguments{
\item{data}{The tibble or data frame containing the data.}

\item{x}{The independant variable (TODO: "variables").}

\item{y}{The dependant variable.}
}
\value{
A dendogram plot.
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
function_name

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
data(iris)
iris_clustering = auto_clustering(iris, x = "Sepal.Length", y = "Sepal.Width")

}
