% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_PCA.R
\name{auto_PCA}
\alias{auto_PCA}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
auto_PCA(data)
}
\arguments{
\item{data}{The tibble or data frame containing the data.}
}
\value{
A list of three elements:
\itemize{
\item \code{loadings}.
\item \code{loadings_plot}.
\item \code{var_explained_plot}.
}
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
function_name

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
iris_pca = auto_PCA(iris)
iris_pca$loadings
iris_pca$loadings_plot
iris_pca$var_explained_plot

}
