% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tree.R
\name{auto_tree}
\alias{auto_tree}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
auto_tree(data, formula, model_mode = "classification", model_depth = 3)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
The fit engine for a tree model, the result of applying \code{extract_fit_engine()} to the trained model.
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
auto_tree

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
data = iris
formula = Species ~ .
auto_tree(data, formula)

}
