% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
check_path(path, print_ok = FALSE)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
check_path

Check if a given path (file or directory) exists
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
