% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric.R
\name{get_numeric}
\alias{get_numeric}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
get_numeric(data, verbose = 1)
}
\arguments{
\item{data}{The tibble or data frame containing the data.}

\item{...}{\code{verbose} for now.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
function_name

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
data(iris)
get_numeric(iris)

}
