% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme.R
\name{ggplot2_theme}
\alias{ggplot2_theme}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
ggplot2_theme()
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
function_name

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
