% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance}
\alias{glance}
\title{This function is used to print information about a given parameter}
\usage{
glance(p_name)
}
\arguments{
\item{p_name}{The name of the parameter}
}
\value{
Nothing, only prints.
}
\description{
glance
}
\details{
This function is used to print information about a given parameter.
}
\examples{
x = 1
glance("x")

}
