% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_environment.R
\name{inspect_environment}
\alias{inspect_environment}
\title{Inspect the environment.}
\usage{
inspect_environment(...)
}
\arguments{
\item{...}{Slaanesh knows what.}
}
\value{
Nothing, only prints.
}
\description{
Inspect the parameters in the working environment.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
inspect_environment
}
\examples{
x = 1
vector_of_items = c("item_1", "item_2")
inspect_environment()

}
