% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmls2tibble.R
\name{kmls2tibbles}
\alias{kmls2tibbles}
\title{Read kml files and convert them to tibbles}
\usage{
kmls2tibbles(kml_files)
}
\arguments{
\item{kml_files}{A vector of kml files.}
}
\value{
A tibble with the data from the kml files.
}
\description{
Read kml files and convert them to tibbles

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

It looks for something like the following:

\if{html}{\out{<div class="sourceCode xml">}}\preformatted{<?xml version="1.0" encoding="UTF-8"?>
<kml xmlns="http://www.opengis.net/kml/2.2">
  <Document>
  <name>Ruta 1 - Murillo</name>
  <Placemark>
     <name>Jardines de Murillo (SP)</name>
     <styleUrl>#icon-1899-DB4436-nodesc</styleUrl>
     <Point>
      <coordinates>
        -5.9890187,37.3831802,0
      </coordinates>
     </Point>
  </Placemark>
  (...)
}\if{html}{\out{</div>}}
}
\details{
kmls2tibble

Read kml files and convert them to tibbles
}
\examples{
print("TODO:")

}
