% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRT.R
\name{MRT}
\alias{MRT}
\title{Get the Mean Radiant Temperature (MRT)}
\usage{
MRT(T_a, GT, V, E = 0.95, D = 0.05, units = "C")
}
\arguments{
\item{T_a}{Air temperature (ºC).}

\item{GT}{Globe temperature (ºC).}

\item{V}{Wind speed (m/s).}

\item{E}{Emissivity of the goble. Defaults to 0.95.}

\item{D}{Diameter of the globe. Defaults to 0.05}

\item{units}{Should the output be given in Kelvin (K) or centigrades (C) degrees? Defaults to centigrades.}
}
\value{
Mean Radiant Temperature
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
MRT

Get Mean Radiant Temperature
}
\examples{
GT = 40
T_a = 40
V = 2

RT(GT, T_a, V)

}
