% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_hist.R
\name{multi_hist}
\alias{multi_hist}
\title{Print histograms for the numeric variables in a tibble.}
\usage{
multi_hist(data, binwidth = "scott")
}
\arguments{
\item{data}{The dataset.}

\item{binwidth}{One numeric value or one of "scott", "fixed", "auto" or "freedman". Defaults to "scott".}
}
\value{
It returns...
}
\description{
Each of the histograms will appear separated.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
multi_hist
}
\section{Warning}{

This is a warning
}

\examples{
multi_hist(iris)

}
