% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outliers.R
\name{multi_outliers}
\alias{multi_outliers}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
multi_outliers(data, plot = "all")
}
\arguments{
\item{data}{The dataset.}

\item{plot}{Do you want to plot the results? Admits "all", "hist", "boxplot".}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \strong{except for the boxplots}. Try \code{multi_hist()} instead for the histograms.
}
\details{
multi_outliers

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
multi_outliers(iris)

}
