% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVR.R
\name{MVR}
\alias{MVR}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
MVR(data)
}
\arguments{
\item{data}{The tibble or data.frame you are working with.}
}
\value{
A tibble with a column for variable names, and another for the MVR, the Missing Value Ratio.
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
MVR

Missing Value Ratio out of a tibble.
}
\section{Warning}{

This is a warning
}

\examples{
set.seed(1974)

iris_pierced <- iris
index <- sample(1:nrow(iris), 10, replace = FALSE)
iris_pierced[index, ] <- NA

MVR(iris_pierced)

}
