% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\name{package_info}
\alias{package_info}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
package_info(only_date = FALSE)
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
package_info

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
