% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hist.R
\name{print_hist}
\alias{print_hist}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
print_hist(data, x_var, fill_color = "skyblue", title = "")
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
print_hist

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
print_boxplot(iris, "Petal.Width")

}
