% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree.R
\name{print_tree}
\alias{print_tree}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
print_tree(tree_model)
}
\arguments{
\item{tree_model}{A tree model, possibly generated by \code{auto_tree()} or the \code{rpart} package.}
}
\value{
Nothing, only prints.
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
print_tree

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
tree_model = auto_tree(iris, Species ~ .)
print_tree(tree_model)

}
