% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
refactor(factor_string, old_names, new_names, ordered = FALSE)
}
\arguments{
\item{factor_string}{A vector with factor structure.}

\item{old_names}{Vector of old names.}

\item{new_names}{Vector of new names.}

\item{ordered}{Do you want it ordered as in \code{new_names}?}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
refactor

Change the values of the levels of a factor.
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
