% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate words using a dictionary}
\usage{
translate(
  words,
  dictionary,
  original_language = "Español",
  translate_to = "English"
)
}
\arguments{
\item{words}{A vector of words to be translated.}

\item{dictionary}{A data frame containing translations. The first column represents the original language,
and the second column represents the target language.}

\item{original_language}{The language of the input words (default is "Español").}

\item{translate_to}{The target language for translation (default is "English").}
}
\value{
A vector of translated words.
}
\description{
This function takes a vector of words, a dictionary, and translates the words from one language to another.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
translate

Translate words from one language to another using a dictionary.
}
\section{Warning}{

Ensure that the dictionary includes the necessary translations for accurate results.
}

\examples{
dictionary <- read.csv2(header=TRUE, text='
  Español;English
  Hola;Hello
  Adiós;Goodbye
  ')

words_to_translate <- c("Hola", "Adiós")
translated_words <- translate(words_to_translate, dictionary)
print(translated_words)
# Output: [1] "Hello"   "Goodbye"

}
