% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTCI.R
\name{UTCI}
\alias{UTCI}
\title{Calculate Universal Thermal Climate Index (UTCI) using a polynomial approximation.}
\usage{
UTCI(ta, tr, vel, rh)
}
\arguments{
\item{ta}{Air temperature (ºC).}

\item{tr}{Mean radiant temperature (ºC).}

\item{vel}{Wind speed 10 m above ground level (m/s).Note that this meteorological speed at 10 m is simply 1.5 times the speed felt at ground in the original Fiala model used to build UTCI.}

\item{rh}{Relative humidity (\%).}
}
\value{
UTCI_approx -- The Universal Thermal Climate Index (UTCI) for the input conditions as approximated by a 4-D polynomial.
}
\description{
This function here is a R version of the Python version of the original UTCI_approx
application written in Fortran.
The original Python code can be found \href{https://github.com/ladybug-tools/ladybug-comfort/blob/e614069d143eb6c8120788bff7fff116dc0d9007/ladybug_comfort/utci.py#LL28}{here}.
The original Fortran code can be found \href{www.utci.org}{here}.
}
\details{
UTCI

Calculate Universal Thermal Climate Index (UTCI) using a polynomial approximation.
}
\note{
[1] Peter Bröde, Dusan Fiala, Krzysztof Blazejczyk, Yoram Epstein,
Ingvar Holmér, Gerd Jendritzky, Bernhard Kampmann, Mark Richards,
Hannu Rintamäki, Avraham Shitzer, George Havenith. 2009.
Calculating UTCI Equivalent Temperature. In: JW Castellani & TL
Endrusick, eds. Proceedings of the 13th International Conference
on Environmental Ergonomics, USARIEM, Natick, MA.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
ta = 24
tr = 24
rh = 60
vel = 1.5

UTCI(ta, tr, vel, rh)

}
