% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTCI2color.R
\name{UTCI2color}
\alias{UTCI2color}
\title{This is a shorter description which appears when the name of the function is recommended.}
\usage{
UTCI2color(UTCI)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
UTCI2color

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
