% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTCI_labels.R
\docType{data}
\name{UTCI_labels}
\alias{UTCI_labels}
\title{This is a shorter description which appears when the name of the function is recommended.}
\format{
An object of class \code{character} of length 11.
}
\usage{
UTCI_labels
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
UTCI_labels

Template for new functions
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
\keyword{datasets}
