% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_hist.R, R/multi_outliers.R
\name{multi_outliers}
\alias{multi_outliers}
\title{Print histograms for the numeric variables in a tibble.}
\usage{
multi_outliers(data, plot = "all")

multi_outliers(data, plot = "all")
}
\arguments{
\item{data}{The dataset.}

\item{plot}{Do you want to plot the results? Admits "all", "hist", "boxplot".}

\item{binwidth}{One numeric value or one of "scott", "fixed", "auto" or "freedman". Defaults to "scott".}
}
\value{
It returns...

It returns...
}
\description{
Each of the histograms will appear separated.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \strong{except for the boxplots}. Try \code{multi_hist()} instead for the histograms.
}
\details{
multi_hist

multi_outliers

Template for new functions
}
\section{Warning}{

This is a warning


This is a warning
}

\examples{
multi_hist(iris)

multi_outliers(iris)

}
